Added feature to load local EH1 and PH1 into the PKHeX Boxes and Editor, fix USB connection

+++ HomePlugin/BotController/ICommunicator.cs
@@ -4,7 +4,7 @@
     {
         void Connect();
         void Disconnect();
-        byte[] ReadBytes(ulong offset, int length);
+        byte[] ReadBytes(uint offset, int length);
         bool Connected { get; set; }
         int Port { get; set; }
         string IP { get; set; }
